Phase-1/requirements.txt
python-dotenv
requests
pandas
SQLAlchemy
psycopg2-binary
gridstatus
pytz
fastapi
uvicorn